LayoutTests/http/tests/cache/partitioned-cache.html

<html>
<head>
<script>
if (window.testRunner) {
	testRunner.dumpAsText();
	testRunner.dumpChildFramesAsText();
	testRunner.waitUntilDone();
	testRunner.setCanOpenWindows(true);
	testRunner.setCloseRemainingWindowsWhenComplete(true);
	internals.settings.setStorageBlockingPolicy('BlockThirdParty');
}

document.location = "http://localhost:8000/cache/resources/partitioned-cache-loader.html";
</script>
<head>
<body>
This test must be run from run-webkit-tests
</body>
</html>
