LayoutTests/platform/mac-mavericks/http/tests/cache/partitioned-cache-expected.txt

CONSOLE MESSAGE: line 10: localhost
CONSOLE MESSAGE: line 10: 127.0.0.1
PASS response is "127.0.0.1"
